Move the logic related to the mouse movement to the mousePosition reducer.
